+++ Dataset/Changes.md
@@ -1,5 +1,6 @@
-# Changes to dataset:
+# Changes from edited.csv to final.csv
 
 * October, 15th, 2017:
   * Deleted 59 rows with empty fields regarding both class of admission and country of citizenship
   * Merged 2 Country of Citizenship columns into 1
+  * Edited SOC code that has year format (11/1/2021 -> 11-1202.01)